+++ src/standards/EthReleaseExponential.sol
@@ -22,8 +22,8 @@ import {
  * Eth Require with Exponential Curve Intent Standard
  * @dev data
  *   [bytes32] standard - the intent standard identifier
- *   [uint48]  startTime - start time of the curve (in seconds)
- *   [uint24]  deltaTime - amount of time from start until curve caps (in seconds)
+ *   [uint40]  startTime - start time of the curve (in seconds)
+ *   [uint32]  deltaTime - amount of time from start until curve caps (in seconds)
  *   [uint96]  startAmount - starting amount
  *   [uint8]   startAmountMult - starting amount multiplier (final_amount = amount * (amountMult * 10))
  *   [uint64]  deltaAmount - amount of change after each second
@@ -85,8 +85,8 @@ contract EthReleaseExponential is IIntentStandard, EthReleaseDelegate {
      */
     function encodeData(
         bytes32 standardId,
-        uint48 startTime,
-        uint24 deltaTime,
+        uint40 startTime,
+        uint32 deltaTime,
         int256 startAmount,
         int256 deltaAmount,
         uint8 exponent,
