+++ src/standards/EthReleaseLinear.sol
@@ -21,8 +21,8 @@ import {
  * Eth Require with Linear Curve Intent Standard
  * @dev data
  *   [bytes32] standard - the intent standard identifier
- *   [uint48]  startTime - start time of the curve (in seconds)
- *   [uint24]  deltaTime - amount of time from start until curve caps (in seconds)
+ *   [uint40]  startTime - start time of the curve (in seconds)
+ *   [uint32]  deltaTime - amount of time from start until curve caps (in seconds)
  *   [uint96]  startAmount - starting amount
  *   [uint8]   startAmountMult - starting amount multiplier (final_amount = amount << amountMult)
  *   [uint64]  deltaAmount - amount of change after each second
@@ -80,7 +80,7 @@ contract EthReleaseLinear is IIntentStandard, EthReleaseDelegate {
      * @param deltaAmount amount of change after each second
      * @return the fully encoded intent standard segment data
      */
-    function encodeData(bytes32 standardId, uint48 startTime, uint24 deltaTime, int256 startAmount, int256 deltaAmount)
+    function encodeData(bytes32 standardId, uint40 startTime, uint32 deltaTime, int256 startAmount, int256 deltaAmount)
         external
         pure
         returns (bytes memory)
