+++ test/foundry/scenarios/TransferEth.sol
@@ -30,8 +30,8 @@ contract TransferEth is ScenarioTestEnvironment {
         UserIntent memory intent = _intent();
         intent = _addErc20ReleaseLinear(
             intent,
-            uint48(block.timestamp - releaseAt),
-            uint24(releaseDuration),
+            uint40(block.timestamp - releaseAt),
+            uint32(releaseDuration),
             releaseStartAmount,
             (releaseEndAmount - releaseStartAmount) / int256(releaseDuration)
         );
